requirements.txt
click>=8.0.0
requests>=2.28.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pandas>=1.5.0
feedparser>=6.0.0
bibtexparser>=1.4.0
pyyaml>=6.0
python-dotenv>=0.19.0
rich>=13.0.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
ollama>=0.1.0
zhipuai>=1.0.0
qianfan>=0.0.5
tqdm>=4.64.0
colorama>=0.4.6
jinja2>=3.1.0
asyncio>=3.4.3
dataclasses-json>=0.5.7